+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Refactored the rounding logic in the treemap chart component
 - Restricted the modification of activity tags in the impersonation mode
 - Hardened the endpoint of the public access for portfolio sharing by restricting it to public accesses
 - Improved the parsing of integer query parameters (`skip` and `take`) in the `GET api/v1/admin/user` endpoint